split up functional tests
